Todos/Findings: true one step for nodes reduce data footprint for nodes -> toLinks to handles with some shifting -> reduce data footprint ##### todos long term: make a boundary for pressure variance and start testing with an object in a fluid investigate the separation of data in multiple points -> separation of wet and dry nodes -> separation of data-array form the nodes -> new class for the macro values rework toLinks, is mostly in order needs more testing thou make wierd channels to should be m4 for me -> next step calculate seems superior than the lookup thou, prob comes down to cpu stalls, should also do a cache hit analysis prob to shed some light on that -> still test out with more tables prob needs a SoA structure to work porperlly thou so more a bigger thing than a smaller... ##### Actual: unsigned vs signed

+++ src/types.h
@@ -21,8 +21,8 @@ extern matrix_t weights;
 
 // structs
 typedef struct coordinate {
-    unsigned x;
-    unsigned y;
+    long x;
+    long y;
 }coordinate_t;
 
 typedef enum nodeIdentifier {
